Changes to Anime, Manga, User Classes To Fix Recursion
Using the anime, manga, and user classes was causing recursion error
messages to show up when parsing other lists. I’m not sure of the cause,
but for now, initialize the values we need to empty arrays instead of
blank objects. Note that this may break clients, so we need to test this.