Add Cookie Login Support
Create a method that will post a login to MAL to grab the login cookies.
This method needs to be called before using fetch or send if the code is
using a front-end page that requires a site login. It is not needed for
direct MAL API calls that use basic auth.

Closes #15