Show Results for Single Item Search Returns
Configure SearchController methods to check for a redirect. If a redirect
occurred, it means only one item matched and the content will be that of
the title’s detail page. In that case, use the appropriate type parser
and return the result. If no redirect occurred, parse as normal.

Fixes #23