Fix ETag on Upcoming, Top, and Search
The ETag currently is set to match the URL of the page, so remove slash
before the query string parameters to make it actually match.

Additionally, for upcoming, add the start_date parameter to ensure the
ETag is actually unique if a request specifies a date. Since we use a
date internally anyway, just always add the start_date parameter to make
it easier and also ensure that the tag changes during a date rollover.