Remove 404 for Empty Search Results
Thinking about it, a 404 isn't really appropriate for empty results. The
URI itself does exist, just the query filter caused no information to be
returned. It's more appropriate to return an empty array indicating no
results than to return a 404 error. This is also how the Ruby API does
it, so it helps compatibility.