Fix "503 Service Unavailable" when added a new friend (after 1 day).
When a new friend is added MAL changes the way it formats stuff.
Example: "3 hours ago" or "Yesterday, g:i A" instead of "m-d-y, g:i A".
This fix checks if the date contains the "-", if not then skip format.

Previous commit fixed the issue of the same day but didn't fixed it the next day.