Establish Versioned Routing
In order for the API to expand, we need to version the endpoints. As a
start, the existing endpoints have been duplicated under a /1/ route
prefix. This will let us then use /1.1/ or any other prefix to mark a
version. For compatibility with existing deployments, the existing
endpoints have also been left at their old location. These will probably
be removed at a later time.