Anime Date Update and Deprecated Element Removal
Update the date in anime so that API2 will return the date as an
ISO8601-formatted date with only the accurate information shown.

Also fix a rare issue with API1 where showing the date for items with a
February date that don't have a defined day return the wrong date when
accessing the information on a day number of 29 or greater where the year
for the item isn't a leap year (or 30 or greater any time).