Extended Personal Anime Details
Extend the controller so that if API 2 or greater is being used and
personal details are being requested, we fetch the extended details page
for anime titles. Build a parser function to parse the new content.
Additionally, extend the Anime model to include the new properties and
annotate so that the serializer will only include them for the correct
API levels.