Fix Dates in Search / Upcoming
Compatibility fix in the parser for search and upcoming. The Ruby API
doesn't set the start or end dates unless they are fully provided on the
page. Adjust the parser to duplicate the behavior. Also, adjust the year
parsing so that it has the right century most of the time.