Anime Personal Data Update Feature
Include new parameters for updating personal details for an anime on the
authenticated user's list. This completely re-works the way the update
system functions. It will only output XML parameters for data that it
has been given so as not to blank or reset other data MAL has. This
change applies to API1 as well.

The only value not supported for updating is the "request discussion"
setting.