API2 Manga Personal Details Error Fix
If requesting personal details for a manga item that hasn't been added
to the user list, there would be errors on API2. This was due to manga
having a different method of access. Now, we only parse extended if
a "listed id" exists, which means the user has it on their list.