Fix Date Helper Month Roll-around
PHP rolls dates around when they don't make sense (April 31st becomes
May 1). When combined with the fact that missing details are filled in
from the current date, this can cause issues when processing date input
with only month and year. Fix the date helper by forcing the day to "1"
when processing the date and then dropping the day in output.