Fix Error in Date Format Helper
In some cases in testing date formats, the logic would drop to the
bottom possibility, but that wasn't correct either. This resulted in a
PHP error. Do an extra check to make sure the createFromFormat() call
didn't return false before trying to format and return.

Additionally, eliminate the last else and move the return to the end of
the function.