Update Connection Test
Switch the connectivity test away from direct socket open to Guzzle. This makes it work more like the API and should work even for configurations that don’t allow fsockopen. Additionally, change to check the connection to MAL so that if that site is down it won’t run the interactive tests.