Recs Uses Right Model
Adjust the recs parser to use the right media type model instead of using the anime model for manga types.

While this works for the output we are doing, it’s an incorrect type internally and has the potential in the future to cause issues if we rely on the model class type.